Add graphical board drawing routine
This routine draws the board using Unicode chess characters while also
coloring in the dark and light squares with the colors cyan and white
respectively.

This routine will here on out be used as the default drawing routine as
it is the most intuitive way of presenting the board to the user.